+++ allennlp/fairness/__init__.py
@@ -6,11 +6,12 @@ This module contains tools to:
 3. debias embeddings during training time and post-processing
 """
 
-from allennlp.fairness.fairness_metrics import (
-    Independence,
-    Separation,
-    Sufficiency,
-    DemographicParityWithoutGroundTruth,
+from allennlp.fairness.fairness_metrics import Independence, Separation, Sufficiency
+from allennlp.fairness.bias_metrics import (
+    WordEmbeddingAssociationTest,
+    EmbeddingCoherenceTest,
+    NaturalLanguageInference,
+    AssociationWithoutGroundTruth,
 )
 from allennlp.fairness.bias_direction import (
     PCABiasDirection,